Dashboard structure (#885)
* refactor: split Calendar

* refactor: split Seed

* refactor: split XP

* refactor: split Leaderboard

* refactor: LatestContent

* fix: rename component

* feat: create GridItem component

* feat: dashboard config with static values

* refactor: implement compact structure with GridItem and utils/dashboard
